+++ R/generator.R
@@ -132,7 +132,12 @@ generator0 <- function(fn, type = "generator") {
     base::local(envir = `_private`, {
       generator_env <- environment()$generator_env
       caller_env <- environment()$caller_env
+
+      # Prevent lints about unknown bindings
       exits <- NULL
+      exited <- NULL
+      cleanup <- NULL
+      close_active_iterators <- NULL
 
       info <- machine_info(type, env = caller_env)
 
@@ -149,8 +154,10 @@ generator0 <- function(fn, type = "generator") {
       env <- new_generator_env(env, info)
       user_env <- env$user_env
 
-      # The compiler caches function bodies, so inline a weak
-      # reference to avoid leaks (#36)
+      # The compiler caches function bodies, so inline a weak reference to avoid
+      # leaks (#36). This weak reference is injected inside the body of the
+      # generator instance to work around a scoping issue. See where we install
+      # the user's exit handlers.
       weak_env <- new_weakref(env)
 
       # Forward arguments inside the user space of the state machine
@@ -159,21 +166,74 @@ generator0 <- function(fn, type = "generator") {
       # Flipped when `f` is pressed in the browser
       undebugged <- FALSE
 
+      # Called on cleanup to close all iterators active in
+      # ongoing `for` loops
+      close_active_iterators <- function() {
+        # The list is ordered from outermost to innermost for loops. Close them
+        # in reverse order, from most nested to least nested.
+        for (iter in rev(env$iterators)) {
+          if (!is_null(iter)) {
+            iter_close(iter)
+          }
+        }
+      }
+
+      env$close_active_iterators <- close_active_iterators
+
+      env$cleanup <- function() {
+        env$close_active_iterators()
+
+        # Prevent user exit handlers from running again
+        env$exits <- NULL
+      }
+
+
       # Create the generator instance. This is a function that resumes
       # a state machine.
-      instance <- inject(function(arg) {
+      instance <- inject(function(arg, close = FALSE) {
         # Forward generator argument inside the state machine environment
         delayedAssign("arg", arg, assign.env = env)
+        delayedAssign("close", close, assign.env = env)
 
         if (!undebugged && (debugged || is_true(peek_option("coro_debug")))) {
           env_browse(user_env)
 
-          on.exit({
+          defer({
             # `f` was pressed, disable debugging for this generator
             if (!env_is_browsed(user_env)) {
               undebugged <<- TRUE
             }
-          }, add = TRUE)
+          })
+        }
+
+        if (is_true(env$exhausted)) {
+          return(exhausted())
+        }
+
+        if (close) {
+          # Prevent returning here as closing should be idempotent. We set
+          # ourselves as exhausted _before_ running any cleanup in case of
+          # failures. An exit handler shouldn't fail and it's expected that any
+          # failure prevents other handlers from running, including when an
+          # attempt is made at resuming the closed generator.
+          env$exhausted <- TRUE
+
+          # First close active iterators. Should be first since they might be
+          # relying on resources set by the user.
+          close_active_iterators()
+
+          # Now run the user's exit expressions. Achieved by running restoring
+          # user exits in the user environment and running an empty eval there.
+          # Unlike in the state machine path, where these expressions are meant
+          # to only run in case of unexpected exits, we don't disable them
+          # before exiting so they will actually run here.
+          evalq(envir = user_env,
+            base::evalq(envir = rlang::wref_key(!!weak_env), {
+              env_poke_exits(user_env, exits)
+            })
+          )
+
+          return(exhausted())
         }
 
         # Disable generator on error, interrupt, debugger quit, etc.
@@ -188,21 +248,20 @@ generator0 <- function(fn, type = "generator") {
           abort("This function has been disabled because of an unexpected exit.")
         }
 
-        if (is_true(env$exhausted)) {
-          return(exhausted())
-        }
-
         # Resume state machine. Set up an execution env in the user
         # environment first to serve as a target for on.exit()
         # expressions. Then evaluate state machine in its private
         # environment.
         env$jumped <- TRUE
-        out <- evalq(envir = user_env,
+        env$exited <- TRUE
+
+        out <- evalq(envir = user_env, {
           base::evalq(envir = rlang::wref_key(!!weak_env), {
+            defer(if (exited) cleanup())
             env_poke_exits(user_env, exits)
             !!state_machine
           })
-        )
+        })
         env$jumped <- FALSE
 
         out
@@ -234,6 +293,7 @@ new_generator_env <- function(parent, info) {
   env$iterators <- list()
   env$handlers <- list()
   env$exits <- NULL
+  env$exited <- TRUE
   env$.last_value <- NULL
 
   with(env, {
@@ -245,6 +305,7 @@ new_generator_env <- function(parent, info) {
     }
 
     suspend <- function() {
+      exited <<- FALSE
       exits <<- env_poke_exits(user_env, NULL)
     }
   })